Rename to Folly
This renames the "new" code to Folly;
the name Encrusted is now only used to refer to
the code that traces back to the original Encrusted repo.

Why rename?
The current code is now quite far from the original implementation,
and the UI has grown from a couple hundred lines of shim code
to quite a large application on its own.
I haven't been able to reach the original Encrusted author,
and I'm hesitant to reuse the name for a bunch of my own code
without permission.